+++ src/main/java/com/planbase/pdf/layoutmanager/pages/RenderTarget.kt
@@ -56,12 +56,12 @@ interface RenderTarget {
      * @param lineStyle the style to draw the line with
      * @return the updated RenderTarget (may be changed to return the lowest y-value instead)
      */
-    @Suppress("unused")
     fun drawLine(start: Coord, end: Coord, lineStyle: LineStyle): IntRange
             = drawLine(start, end, lineStyle, true)
 
     /**
-     * Draws lines from the first point to the last.  Direction is important if using mitering.
+     * Draws lines from the first point to the last using the given line style.  PDF has only one line style for
+     * a line strip.
      *
      * @param points the list of Coord to draw lines between.  This does *not* connect the last point to the first.
      * If you want that, add the first point again at the end of the list.
@@ -72,16 +72,24 @@ interface RenderTarget {
     fun drawLineStrip(points: List<Coord>, lineStyle: LineStyle, reallyRender: Boolean = true): IntRange
 
     /**
-     * Draws lines from the first point to the last.  Direction is important if using mitering.
-     *
-     * @param points the list of Coord to draw lines between.  This does *not* connect the last point to the first.
-     * If you want that, add the first point again at the end of the list.
-     * @param lineStyle the style to draw the line with
-     * @return the updated RenderTarget (may be changed to return the lowest y-value instead)
+     * Draws lines from the first point to the last.  Convenience function for [drawLineStrip(points, lineStyle, true)]
      */
-    @Suppress("unused")
     fun drawLineStrip(points: List<Coord>, lineStyle: LineStyle): IntRange =
             drawLineStrip(points, lineStyle, true)
+
+    /**
+     * Draws a closed path.
+     *
+     * @param points the list of Coord to draw lines between.  The last point is assumed to connect back to the first.
+     * @param lineStyle the style to draw all lines with.  PDF only allows one line width and color per path.
+     * @param reallyRender Only render this if true.  Otherwise, just measure and return.
+     * @return the updated RenderTarget (may be changed to return the lowest y-value instead)
+     */
+    fun drawLineLoop(points: List<Coord>, lineStyle: LineStyle, reallyRender: Boolean): IntRange
+
+    /** Draws a closed path.  Convenience function for [drawLineLoop(points, lineStyles, true)]. */
+    fun drawLineLoop(points: List<Coord>, lineStyle: LineStyle): IntRange =
+            drawLineLoop(points, lineStyle, true)
 
     /**
      * Puts styled text on this RenderTarget